# .migration-backup/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * SteamShare API - Steam account marketplace with points, XP, and badges
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface SuccessMessage {
  message: string;
}

export interface RegisterInput {
  /**
     * @minLength 3
     * @maxLength 30
     */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  points: number;
  xp: number;
  level: number;
  /** @nullable */
  badgeName?: string | null;
  isAdmin: boolean;
  isBanned: boolean;
  /** @nullable */
  banReason?: string | null;
  /** @nullable */
  banExpiresAt?: string | null;
  createdAt: string;
}

export interface UserPublic {
  id: number;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  points: number;
  xp: number;
  level: number;
  /** @nullable */
  badgeName?: string | null;
  isBanned?: boolean;
  createdAt: string;
}

export interface UserProfile {
  id: number;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  points: number;
  xp: number;
  level: number;
  /** @nullable */
  badgeName?: string | null;
  totalAccounts: number;
  totalLikesReceived: number;
  isBanned?: boolean;
  createdAt: string;
}

export type AccountUnlockMethod = typeof AccountUnlockMethod[keyof typeof AccountUnlockMethod];


export const AccountUnlockMethod = {
  login: 'login',
  like: 'like',
  comment: 'comment',
} as const;

export interface Account {
  id: number;
  userId: number;
  username: string;
  posterUsername?: string;
  /** @nullable */
  posterAvatarUrl?: string | null;
  title: string;
  description: string;
  games: string[];
  pointsCost: number;
  isAvailable: boolean;
  likesCount: number;
  claimsCount: number;
  viewCount?: number;
  workingVotes?: number;
  notWorkingVotes?: number;
  unlockMethod?: AccountUnlockMethod;
  userHasLiked?: boolean;
  userHasCommented?: boolean;
  createdAt: string;
}

export interface AccountListResponse {
  accounts: Account[];
  total: number;
  page: number;
  limit: number;
}

export type AccountInputUnlockMethod = typeof AccountInputUnlockMethod[keyof typeof AccountInputUnlockMethod];


export const AccountInputUnlockMethod = {
  login: 'login',
  like: 'like',
  comment: 'comment',
} as const;

export interface AccountInput {
  /**
     * @minLength 3
     * @maxLength 100
     */
  title: string;
  /** @maxLength 1000 */
  description: string;
  /** @minItems 1 */
  games: string[];
  /** @minimum 0 */
  pointsCost: number;
  steamUsername: string;
  steamPassword: string;
  unlockMethod?: AccountInputUnlockMethod;
}

export interface ClaimResult {
  steamUsername: string;
  steamPassword: string;
  pointsSpent: number;
}

export interface Comment {
  id: number;
  accountId: number;
  userId: number;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  content: string;
  likesCount: number;
  userHasLiked?: boolean;
  createdAt: string;
}

export interface CommentInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  content: string;
  parentId?: number | null;
}

export interface GameCount {
  game: string;
  count: number;
}

export interface AdLink {
  id: number;
  code: string;
  /** @nullable */
  description?: string | null;
  pointsReward: number;
  maxUses: number;
  usesCount: number;
  isActive: boolean;
  createdAt: string;
}

export interface AdLinkInput {
  description?: string;
  /** @minimum 1 */
  pointsReward: number;
  /** @minimum 1 */
  maxUses: number;
}

export interface RedeemResult {
  pointsEarned: number;
  newTotal: number;
  xpEarned?: number;
}

export interface Badge {
  id: number;
  name: string;
  description: string;
  xpThreshold: number;
  /** @nullable */
  iconUrl?: string | null;
}

export interface SiteStats {
  totalUsers: number;
  totalAccounts: number;
  totalClaims: number;
  totalPointsCirculating: number;
  topGames: GameCount[];
}

export interface ForgotPasswordInput {
  email: string;
}

export interface ForgotPasswordResult {
  message: string;
  resetToken?: string;
  expiresAt?: string;
}

export interface ResetPasswordInput {
  token: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface VerifyCredentialsInput {
  steamUsername: string;
  steamPassword: string;
}

export type VerifyCredentialsResultStatus = typeof VerifyCredentialsResultStatus[keyof typeof VerifyCredentialsResultStatus];


export const VerifyCredentialsResultStatus = {
  valid: 'valid',
  invalid: 'invalid',
  rate_limited: 'rate_limited',
  error: 'error',
} as const;

export interface VerifyCredentialsResult {
  status: VerifyCredentialsResultStatus;
  message: string;
  games?: string[];
  steamid?: string;
  isFamilyShare?: boolean;
}

export interface Giveaway {
  id: number;
  createdBy: number;
  title: string;
  description: string;
  prize: string;
  taskDescription: string;
  maxEntries: number;
  entriesCount: number;
  endDate: string;
  isActive: boolean;
  /** @nullable */
  winnerUserId?: number | null;
  /** @nullable */
  winnerUsername?: string | null;
  userHasEntered?: boolean;
  createdAt: string;
}

export interface GiveawayInput {
  title: string;
  description: string;
  prize: string;
  taskDescription: string;
  maxEntries?: number;
  endDate: string;
}

export interface EnterGiveawayInput {
  taskProof?: string;
}

export interface GiveawayEntry {
  id: number;
  giveawayId: number;
  userId: number;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  taskProof?: string | null;
  createdAt: string;
}

export interface DrawResult {
  message: string;
  winnerUserId?: number;
  /** @nullable */
  winnerUsername?: string | null;
}

export type ListAccountsParams = {
game?: string;
sort?: ListAccountsSort;
page?: number;
limit?: number;
};

export type ListAccountsSort = typeof ListAccountsSort[keyof typeof ListAccountsSort];


export const ListAccountsSort = {
  recent: 'recent',
  popular: 'popular',
  free: 'free',
  points: 'points',
} as const;

export type GetLeaderboardParams = {
limit?: number;
};

export type ListUsersParams = {
page?: number;
};