# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SteamShare API - Steam account marketplace with points, XP, and badges
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts/{accountId}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get a single steam account listing
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete a steam account listing (owner or admin)
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /accounts/{accountId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments on an account listing
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

    post:
      operationId: createComment
      tags: [comments]
      summary: Post a comment on an account listing
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /accounts/{accountId}/comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment (owner or admin)
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /accounts/{accountId}/comments/{commentId}/like:
    post:
      operationId: likeComment
      tags: [comments]
      summary: Like a comment
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

    delete:
      operationId: unlikeComment
      tags: [comments]
      summary: Unlike a comment
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unliked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  # Users
...
  /ad-links:
    get:
      operationId: listAdLinks
      tags: [adlinks]
      summary: List all ad links (admin only)
      responses:
        "200":
          description: Ad links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdLink"

    post:
      operationId: createAdLink
      tags: [adlinks]
      summary: Create an ad link (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdLinkInput"
      responses:
        "201":
          description: Ad link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdLink"
  /ad-links/{code}/redeem:
    get:
      operationId: redeemAdLink
      tags: [adlinks]
      summary: Redeem an ad link to earn points
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Points earned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedeemResult"
        "400":
          description: Already redeemed or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Link not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /ad-links/{adLinkId}:
    delete:
      operationId: deleteAdLink
      tags: [adlinks]
      summary: Delete an ad link (admin only)
      parameters:
        - name: adLinkId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  # Admin
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Account:
      type: object
      required: [id, userId, username, title, description, games, pointsCost, isAvailable, likesCount, claimsCount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        posterUsername:
          type: string
        posterAvatarUrl:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        games:
          type: array
          items:
            type: string
        pointsCost:
          type: integer
        isAvailable:
          type: boolean
        likesCount:
          type: integer
        claimsCount:
          type: integer
        viewCount:
          type: integer
        workingVotes:
          type: integer
        notWorkingVotes:
          type: integer
        unlockMethod:
          type: string
          enum: [login, like, comment]
        userHasLiked:
          type: boolean
        userHasCommented:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Comment:
      type: object
      required: [id, accountId, userId, username, content, likesCount, createdAt]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        content:
          type: string
        likesCount:
          type: integer
        userHasLiked:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 500
        parentId:
          type: ["integer", "null"]
...
    AdLink:
      type: object
      required: [id, code, pointsReward, maxUses, usesCount, isActive, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        description:
          type: ["string", "null"]
        pointsReward:
          type: integer
        maxUses:
          type: integer
        usesCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AdLinkInput:
      type: object
      required: [pointsReward, maxUses]
      properties:
        description:
          type: string
        pointsReward:
          type: integer
          minimum: 1
        maxUses:
          type: integer
          minimum: 1
    RedeemResult:
      type: object
      required: [pointsEarned, newTotal]
      properties:
        pointsEarned:
          type: integer
        newTotal:
          type: integer
        xpEarned:
          type: integer
...